test(tuttid): split daemon API route tests
